Skip single test causing new failures
Last-Update: Sun, 18 Feb 2018 12:06:22 +0200
Bug-Debian: https://bugs.debian.org/890704
Bug-Upstream: https://github.com/samtools/bcftools/issues/744

This is a new test that was added in 1.7, so it is not a regression.

One option is to simply skip it for now.

FIXME: This workaround needs to be reconsidered in future versions

Gbp-Pq: Name skip_new_test_failing_on_32bit.patch