Set the bcftools plugin path to search system directories
Forwarded: not-needed
Last-Update: 2015-11-09

The tests were intended to be run at build time only, so the test
script sets the plugins to be searched for /only/ in the build directory.
It does so by setting the environment variable BCFTOOLS_PLUGINS. However,
as described in the manpage bcftools(1), adding a colon to either side of
the path entry in the environment variable also searches the system
directories. We need this for autopkgtest to work.
===================================================================

Gbp-Pq: Name tests-pluginpath.patch